+++ app/src/gui/pages/record-create.tsx
@@ -57,7 +57,7 @@ import {ParentLinkProps} from '../components/record/relationships/types';
 import DraftSyncStatus from '../components/record/sync_status';
 import BackButton from '../components/ui/BackButton';
 import Breadcrumbs from '../components/ui/breadcrumbs';
-import {checkIfParentHasInheritedData} from '../../utils/checkParentInheritedData';
+import {checkIfParentHasInheritedData} from '../../utils/formUtilities';
 
 interface DraftCreateActionProps {
   project_id: ProjectID;
